+++ docs/default.nix
@@ -9,7 +9,16 @@ stdenv.mkDerivation {
 
   src = ./.;
 
-  buildInputs = [ pandoc python39Packages.weasyprint ];
+  buildInputs =
+    let
+      # Workaround to allow setting custom fontconfig file.
+      # Should be fixed in next version of Nixpkgs, so the regular package
+      # can again be used (https://github.com/NixOS/nixpkgs/pull/254239).
+      weasyprint = python39Packages.weasyprint.overrideAttrs (o: {
+        makeWrapperArgs = [ "--set-default FONTCONFIG_FILE ${o.FONTCONFIG_FILE}" ];
+      });
+    in
+    [ pandoc weasyprint ];
 
   installPhase = ''
     DATE=$(date -I)