+++ live/default.nix
@@ -1,39 +1,45 @@
 # Build NixOS system
-{pkgs, lib, version, updateCert, kioskUrl, greeting, playos-controller}:
+{pkgs, lib, kioskUrl, playos-controller, application}:
 with lib;
 let nixos = pkgs.importFromNixos ""; in
 (nixos {
   configuration = {...}: {
     imports = [
-      # general PlayOS modules
-      ((import ../system/modules/playos.nix) {inherit pkgs version updateCert kioskUrl greeting playos-controller;})
-      # Play Kiosk and Driver
-      ../system/play-kiosk.nix
-      # Networking
-      ../system/networking
-      # Localization
-      ../system/localization.nix
+      # Base layer
+      ((import ../base) {
+        inherit pkgs kioskUrl playos-controller;
+        inherit (application) safeProductName fullProductName greeting version;
+      })
+      # Application-specific
+      application.module
 
       (pkgs.importFromNixos "modules/installer/cd-dvd/iso-image.nix")
     ];
 
     config = {
-      # Force use of already overlayed nixpkgs in modules
-      nixpkgs.pkgs = pkgs;
-
       # ISO image customization
       isoImage.makeEfiBootable = true;
       isoImage.makeUsbBootable = true;
-      isoImage.isoName = "playos-live-${version}.iso";
+      isoImage.isoName = "${application.safeProductName}-live-${application.version}.iso";
       isoImage.appendToMenuLabel = " Live System";
 
       # Set up as completely volatile system
-      systemPartition.enable = mkForce false;
-      volatileRoot.persistentDataPartition = {
+      fileSystems."/boot" = {
         device = "tmpfs";
         fsType = "tmpfs";
         options = [ "mode=0755" ];
       };
+      playos.storage = {
+        systemPartition.enable = false;
+        persistentDataPartition = {
+          device = "tmpfs";
+          fsType = "tmpfs";
+          options = [ "mode=0755" ];
+        };
+      };
+
+      # Live system does not self update
+      playos.selfUpdate.enable = false;
 
       # There is no persistent state for a live system
       system.stateVersion = lib.trivial.release;