+++ pkgs/default.nix
@@ -1,11 +1,11 @@
-{ version, updateUrl, kioskUrl, activeVirtualTerminals ? [] }:
+{ applicationOverlays ? [] }:
 
 let
 
   nixpkgs = builtins.fetchTarball {
-    # nixos-22.11 2023-01-11
-    url = "https://github.com/nixos/nixpkgs/archive/54644f409ab471e87014bb305eac8c50190bcf48.tar.gz";
-    sha256 = "1pqgwbprmm84nsylp8jjhrwchzn3cv9iiaz1r89mazfil9qcadz0";
+    # nixos-23.05 2023-08-14
+    url = "https://github.com/nixos/nixpkgs/archive/720e61ed8de116eec48d6baea1d54469b536b985.tar.gz";
+    sha256 = "0ii10wmm8hqdp7bii7iza58rjaqs4z3ivv71qyix3qawwxx48hw9";
   };
 
   overlay =
@@ -15,39 +15,17 @@ let
 
       rauc = (import ./rauc) super;
 
-      dividat-driver = (import ./dividat-driver) {
-        inherit (super) stdenv fetchFromGitHub buildGoModule;
-        pkgs = self;
-      };
-
-      playos-kiosk-browser = import ../kiosk {
-        pkgs = self;
-        system_name = "PlayOS";
-        system_version = version;
-      };
-
-      breeze-contrast-cursor-theme = super.callPackage ./breeze-contrast-cursor-theme {};
-
       ocamlPackages = super.ocamlPackages.overrideScope' (self: super: {
         semver = self.callPackage ./ocaml-modules/semver {};
         obus = self.callPackage ./ocaml-modules/obus {};
+        opium = self.callPackage ./ocaml-modules/opium {};
+        opium_kernel = self.callPackage ./ocaml-modules/opium_kernel {};
       });
-
-      # Controller
-      playos-controller = import ../controller {
-        pkgs = self;
-        version = version;
-        updateUrl = updateUrl;
-        kioskUrl = kioskUrl;
-      };
 
     };
 
 in
 
-  import nixpkgs {
-    overlays = [
-      overlay
-      (import ./xorg { inherit activeVirtualTerminals; })
-    ];
-  }
+import nixpkgs {
+  overlays = [ overlay ] ++ applicationOverlays;
+}